Implement Winner Announcing
Now when a game ends, the winner (and loser) is announced :). Not only
that, the permissions of respective players and @everyone is changed
to view only.
